Add SMTP adapter
This is the first version of an SMTP adapter backed by gen_stmp.
It only supports text/plain, text/html, and multipart/alternative parts
with limited support for configuring the Content-Type,
Content-Transfer-Encoding of each individual parts.

There is also no support for attachments since Swoosh doesn't support it
yet.

The testing is not as good as I would like it to be, ideally I would
want to test whatever the mimemail modules returns (the string that will
be sent to the server), rather than the tuple I'm generating.