Support passing config as paramater to deliver/2
For some providers it's useful to be able to override the Mailer
config on a per-email basis. For instance, Mailgun requires you
to specify the domain you are sending from. If you are building
an app that's supporting several domains then you can use that
feature to pass the domain when making the call to deliver/2.